- Dine In Floor Plan Data Types changes,Select Order Type Dialog Crash Solved

+++ app/src/main/java/com/android/pos/ui/fragments/dinein/DineInFragment.kt
@@ -155,8 +155,8 @@ class DineInFragment : Fragment() {
                         inflatedViewSquare.tag = dineInFloorTablesList[i]
 
                         val paramsSquare = FrameLayout.LayoutParams(
-                            dineInFloorTablesList[i].width.toDp(),
-                            dineInFloorTablesList[i].height.toDp()
+                            dineInFloorTablesList[i].width.toInt().toDp(),
+                            dineInFloorTablesList[i].height.toInt().toDp()
                         )
 
                         paramsSquare.leftMargin = dineInFloorTablesList[i].xPosition.toInt().toDp()
@@ -199,8 +199,8 @@ class DineInFragment : Fragment() {
                         inflatedViewRound.tag = dineInFloorTablesList[i]
 
                         val paramsRound = FrameLayout.LayoutParams(
-                            dineInFloorTablesList[i].width.toDp(),
-                            dineInFloorTablesList[i].height.toDp()
+                            dineInFloorTablesList[i].width.toInt().toDp(),
+                            dineInFloorTablesList[i].height.toInt().toDp()
                         )
                         paramsRound.leftMargin = dineInFloorTablesList[i].xPosition.toInt().toDp()
                         paramsRound.topMargin = dineInFloorTablesList[i].yPosition.toInt().toDp()
